added func

+++ UserInput.cs
@@ -3,10 +3,10 @@ using System.Text.Json.Serialization;
 public class UserInput
 {
     [JsonPropertyName("lat")]
-    public float lat { get; set; }
+    public double lat { get; set; }
 
     [JsonPropertyName("lng")]
-    public float lng { get; set; }
+    public double lng { get; set; }
 
     [JsonPropertyName("radius")]
     public int radius { get; set; }
